refactor: Candle 가격 필드 BigDecimal 전환 및 CandleInterval duration 추가 - Candle OHLC 필드 double → BigDecimal 변경 - CandleInterval에 Duration 필드 추가 (동적 range 계산용) - CandleFilter VO 추가 (포트 파라미터 캡슐화)

+++ src/main/java/ksh/tryptobackend/marketdata/domain/model/Candle.java
@@ -1,12 +1,13 @@
 package ksh.tryptobackend.marketdata.domain.model;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public record Candle(
     Instant time,
-    double open,
-    double high,
-    double low,
-    double close
+    BigDecimal open,
+    BigDecimal high,
+    BigDecimal low,
+    BigDecimal close
 ) {
 }
